Configure the SQL file, set up the database content I need in the file, open phpMyAdmin and input the SQL command to generate the database, and create a new server.js file related to the API server. During this process, install the dependencies and test the server.
I originally planned to keep the database in Chinese temporarily for my own convenience when checking, but this would make subsequent modifications too complicated. Therefore, I have now changed everything to English. However, the comments in the code are still in Chinese to help me quickly locate and make changes.